# mnemotopia-back/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "My First API",
        "version": "0.1"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api/",
            "description": "Developing env"
        }
    ],
    "paths": {
        "/chain-elements": {
            "get": {
                "tags": [
                    "chain-elements"
                ],
                "operationId": "b897e53bce54e69df4dd68520e26c2b8",
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "chain-elements"
                ],
                "operationId": "c26f2045b9083fbf985b93e0d381971f",
                "parameters": [
                    {
                        "name": "chain_id",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/chain-elements/{chain-element}": {
            "get": {
                "tags": [
                    "chain-elements"
                ],
                "operationId": "edb6e1c27f0d4ec0913aca372951925d",
                "parameters": [
                    {
                        "name": "chain-element",
                        "in": "path",
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "chain-elements"
                ],
                "operationId": "56a857276779946bc89afec56bb36b77",
                "parameters": [
                    {
                        "name": "chain-element",
                        "in": "path",
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "patch": {
                "tags": [
                    "chain-elements"
                ],
                "operationId": "61c0c5e65e844d16fa354118f905e5b9",
                "parameters": [
                    {
                        "name": "chain-element",
                        "in": "path",
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "chain_id": {
                                        "type": "integer"
                                    },
                                    "order": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/chains": {
            "get": {
                "tags": [
                    "chains"
                ],
                "operationId": "011c3e81f0b7efa70a0f00d337aa070a",
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "chains"
                ],
                "operationId": "2221d762f0774d03bbf61df2a0a2cc53",
                "parameters": [
                    {
                        "name": "text",
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "group_id",
                        "in": "query",
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/chains/{chain}": {
            "get": {
                "tags": [
                    "chains"
                ],
                "operationId": "00b4529499bc798798a0f29770aa7f4f",
                "parameters": [
                    {
                        "name": "chain",
                        "in": "path",
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "chains"
                ],
                "operationId": "235ca6b7441f2ee681b2423955568cfc",
                "parameters": [
                    {
                        "name": "chain",
                        "in": "path",
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "patch": {
                "tags": [
                    "chains"
                ],
                "operationId": "df39736372d14e47c78f311969ac3246",
                "parameters": [
                    {
                        "name": "chain",
                        "in": "path",
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "group_id": {
                                        "type": "integer"
                                    },
                                    "text": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    }
}